nx-vpu: match bufer by dma addr; possibility to request additional buffers

+++ drivers/media/platform/nxp-vpu/nx_vpu_api.c
@@ -521,8 +521,8 @@ int NX_VpuInit(void *dev, void *baseAddr, void *firmVirAddr,
 	for (i = 0 ; i < NX_MAX_VPU_INSTANCE ; i++) {
 		gstVpuInstance[i].inUse = 0;
 		gstVpuInstance[i].paramPhyAddr = paramBufAddr;
-		gstVpuInstance[i].paramVirAddr = (unsigned long)(firmVirAddr +
-			CODE_BUF_SIZE + TEMP_BUF_SIZE);
+		gstVpuInstance[i].paramVirAddr = firmVirAddr +
+			CODE_BUF_SIZE + TEMP_BUF_SIZE;
 		gstVpuInstance[i].paramBufSize = PARA_BUF_SIZE;
 	}
 	gstIsInitialized = 1;
